requirements.txt
# Core Flask application dependencies
Flask>=2.3.2
Werkzeug>=2.3.6

# Environment variables and API key handling
python-dotenv>=1.0.0

# OpenAI API, embeddings, and language models
openai>=0.28.0
langchain>=0.0.292
langchain_community>=0.1.0
langchain_openai>=0.1.0

# Embedding and tokenization libraries
tiktoken>=0.4.0
tenacity>=8.0.1

# PDF file handling and processing
pdfplumber>=0.6.0
PyPDF2>=3.0.1

# Vector search and storage (use faiss-gpu if available)
faiss-cpu>=1.7.4

# HTTP request handling and scraping utilities
httpx>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
cloudscraper>=1.2.69

# Logging and debugging utilities (optional, included for clarity)
# Logging is part of the standard library, so no installation is required

# Optional utilities for better development experience
# tqdm (for progress bars, optional)
# gunicorn (for production deployment of Flask apps, optional)
# pytz (for timezone handling, optional)

# Utility libraries for JSON handling and other miscellaneous tasks
json5>=0.9.11
